Changed how the VERSION property is read/written:
Parsing:
* If the VERSION property contains a recognized value (i.e. "1.0" or
"2.0"), then it will NOT be unmarshalled as a "Version" property object.
Instead, the "ICalendar" object will be assigned the appropriate
"ICalVersion" enum value.
* If the VERSION property contains an unrecognized value, then it will
be unmarshalled as a "Version" property object as before.

Writing:
* If the "ICalendar" object does not contain a "Version" property
object, then a VERSION property will be inserted automatically, based on
the writer's target version.
* If the "ICalendar" object contains a "Version" property, then it will
be written like any other property, as before.