Fixed a date-time property bug (see Ticket 7).
If a DTSTART property (or any other date-time property) with a DATE
value is read and then immediately written, the written property
incorrectly has a date-time value.  This is because the scribe was
checking for the existance of raw components first before checking for
the existance of a date object.