Resolve Issue #49 in BiWeekly
The logic for converting to UTC from a local time and from UTC to a local time in TimeUtils.convert method has been simplified. Once the milliseconds from the Epoch has been obtained, those milliseconds and the relevant timezone are passed to method "toDateTimeValue(long millisFromEpoch, TimeZone zone)" in TimeUtils.java for conversion to a DateTimeValue type. That method converts the milliseconds from the Epoch into a GregorianCalendar type first. This steps takes care of all timezone offsets adjustments as well as handling the DST "gap" and "overlap" hours when applicable.