Fix issue with UNTIL value of a recurrence not being handled correctly
The unit tests were failing under certain timezones because of the way
the UNTIL date object was being constructed in the unit tests. It was
constructed using the SimpleDateFormat#parse method. By default, this
method assumes the given date string is under the default timezone,
which caused a different UNTIL timestamp to be created for different
people around the world. The recurrence iterator code wasn't taking this
into account, so I modified it to behave differently if the given UNTIL
value lacks a time component (the user must announce this by passing
"false" into the second parameter of the ICalDate constructor).

The other problem was that it was chopping off the time components of
the UNTIL value if the start date lacked time components, which, under
certain timezones, essentially caused the UNTIL value to be decremented
by one day, which is why the last date in a recurrence was not being
returned for so many people. See pseudocode below.

until = SimpleDateFormat.parse("2006-05-03") //May 3 2006 00:00
America/New_York
until = convertToUtc(until) //May 2 2006 20:00 UTC
until = removeTimeComponents(until) //May 2 2006

https://github.com/mangstadt/biweekly/pull/63
https://github.com/mangstadt/biweekly/issues/67
https://github.com/mangstadt/biweekly/issues/74
https://github.com/mangstadt/biweekly/issues/99