Fixed issue with "advanceTo" not working correctly
For what I assume to be performance reasons, the advanceTo method
contains code that "jumps" the iterator ahead a certain amount, so that
it does not have to iterate over every single date in the recurrence
until it reaches the stop date. As you suggested, this code was not
taking the day of the month into account. Setting the day of the month
to 1 seems to solve the issue.

https://github.com/mangstadt/biweekly/issues/69