Fix problem with recurrences that use BYSETPOS
If (a) a recurrence contained BYSETPOS, and (b) the start date of the
recurrence matched the first date in the recurrence set, it was skipping
this date and starting with the next date in the recurrence set.

This was happening because the time component of the start date was
being zeroed-out.

https://github.com/mangstadt/biweekly/issues/122